Вынесли логику загрузки данных из DetailsFragment в отдельный класс WeatherLoader.kt
